separated string storage
